rename package to cli
