+++ .gitignore
@@ -1,4 +1,5 @@
 appsdeck-cli
+bin
 pkg
 src/github.com
 src/code.google.com
