+++ src/appsdeck/auth/destroy.go
@@ -1,0 +1,10 @@
+package auth
+
+import (
+	"appsdeck/constants"
+	"os"
+)
+
+func DestroyToken() error {
+	return os.Remove(constants.AuthConfigFile)
+}
