--- src/appsdeck/cli/api/apps.go
@@ -1,21 +1,0 @@
-package api
-
-import (
-	"net/http"
-)
-
-func AppsList() (*http.Response, error) {
-	req := map[string]interface{}{
-		"method":   "GET",
-		"endpoint": "/api/apps",
-	}
-	return Do(req)
-}
-
-func AppsShow(app string) (*http.Response, error) {
-	req := map[string]interface{}{
-		"method":   "GET",
-		"endpoint": "/api/apps/" + app,
-	}
-	return Do(req)
-}
