--- src/appsdeck/cli/auth/destroy.go
@@ -1,10 +1,0 @@
-package auth
-
-import (
-	"appsdeck/cli/constants"
-	"os"
-)
-
-func DestroyToken() error {
-	return os.Remove(constants.AuthConfigFile)
-}
