--- src/appsdeck/cli/cmd/error.go
@@ -1,11 +1,0 @@
-package cmd
-
-import (
-	"fmt"
-	"os"
-)
-
-func errorQuit(err error) {
-	fmt.Printf("[Error] %s\n", err)
-	os.Exit(1)
-}
