--- src/appsdeck/cli/constants/auth.go
@@ -1,5 +1,0 @@
-package constants
-
-var (
-	TokenName = "auth_token"
-)
