--- src/appsdeck/cli/debug/log.go
@@ -1,28 +1,0 @@
-package debug
-
-import (
-	"log"
-	"os"
-)
-
-var (
-	Enable bool
-)
-
-func init() {
-	if os.Getenv("DEBUG") == "1" {
-		Enable = true
-	}
-}
-
-func Println(vars ...interface{}) {
-	if Enable {
-		log.Println(vars...)
-	}
-}
-
-func Printf(format string, vars ...interface{}) {
-	if Enable {
-		log.Printf(format, vars...)
-	}
-}
