+++ src/appsdeck/cmd/error.go
@@ -1,0 +1,11 @@
+package cmd
+
+import (
+	"fmt"
+	"os"
+)
+
+func errorQuit(err error) {
+	fmt.Printf("[Error] %s\n", err)
+	os.Exit(1)
+}
