+++ src/appsdeck/cmd/logout.go
@@ -1,0 +1,20 @@
+package cmd
+
+import (
+	"appsdeck/auth"
+	"fmt"
+	"github.com/codegangsta/cli"
+)
+
+var (
+	LogoutCommand = cli.Command{
+		Name:  "logout",
+		Usage: "Logout from Appsdeck",
+		Action: func(c *cli.Context) {
+			if err := auth.DestroyToken(); err != nil {
+				panic(err)
+			}
+			fmt.Println("Appsdeck credentials have been deleted.")
+		},
+	}
+)
