+++ src/appsdeck/constants/auth.go
@@ -1,0 +1,5 @@
+package constants
+
+var (
+	TokenName = "auth_token"
+)
